init v2.3.0

+++ components/rtthread/libcpu/riscv/c907fdvm/cpuport_smp.c
@@ -29,7 +29,7 @@ __attribute__((weak)) void riscv_soc_init_cpu(void)
 
 typedef struct _spinlock
 {
-    int lock;
+    long lock;
 } spinlock_t;
 
 #define mb()                          \
